rollback public variables
