+++ src/main/java/codechicken/nei/BookmarkPanel.java
@@ -5,7 +5,7 @@ import codechicken.nei.util.NBTJson;
 import codechicken.nei.recipe.StackInfo;
 import codechicken.nei.recipe.BookmarkRecipeId;
 import codechicken.nei.guihook.GuiContainerManager;
-import codechicken.nei.ItemsGrid.ItemPanelSlot;
+import codechicken.nei.ItemPanel.ItemPanelSlot;
 import com.google.gson.JsonParser;
 import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonObject;
